scripts/skills/requirements.txt
# For bulk_extract_skill_yaml.py (plan / apply validation)
PyYAML>=6
